fix: 새로고침시 userData 받아오지 못하는 에러 해결
